Integration of Firebase! v0.02a
Finally! Integrated the firebase functionalities to the code and now the name, balance and the photo id completely depend on the firebase. Fixed some bugs on loading. Converted all the online links (except images) to offline links. Did more modifications on the Charset. However still getting error on Console: "[Deprecation] Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/." Need to look more into this. However, this was a good update. P.S  - Cleaned asthetics of code, and deleted unused redunencies